+++ typescript/cli/src/avs/config.ts
@@ -16,4 +16,10 @@ export const avsAddresses: ChainMap<AVSContracts> = {
     ecdsaStakeRegistry: '0xFfa913705484C9BAea32Ffe9945BeA099A1DFF72',
     hyperlaneServiceManager: '0xc76E477437065093D353b7d56c81ff54D167B0Ab',
   },
+  ethereum: {
+    avsDirectory: '0x135dda560e946695d6f155dacafc6f1f25c1f5af',
+    proxyAdmin: '0x75EE15Ee1B4A75Fa3e2fDF5DF3253c25599cc659',
+    ecdsaStakeRegistry: '0x272CF0BB70D3B4f79414E0823B426d2EaFd48910',
+    hyperlaneServiceManager: '0xe8E59c6C8B56F2c178f63BCFC4ce5e5e2359c8fc',
+  },
 };