+++ typescript/cli/src/validator/address.ts
@@ -141,7 +141,7 @@ function getEthereumAddress(publicKey: Buffer): string {
 async function getAccessKeyId(skipConfirmation: boolean) {
   if (skipConfirmation) throw new Error('No AWS access key ID set.');
   else
-    return input({
+    return await input({
       message:
         'Please enter AWS access key ID or use the AWS_ACCESS_KEY_ID environment variable.',
     });
@@ -150,7 +150,7 @@ async function getAccessKeyId(skipConfirmation: boolean) {
 async function getSecretAccessKey(skipConfirmation: boolean) {
   if (skipConfirmation) throw new Error('No AWS secret access key set.');
   else
-    return input({
+    return await input({
       message:
         'Please enter AWS secret access key or use the AWS_SECRET_ACCESS_KEY environment variable.',
     });
@@ -159,7 +159,7 @@ async function getSecretAccessKey(skipConfirmation: boolean) {
 async function getRegion(skipConfirmation: boolean) {
   if (skipConfirmation) throw new Error('No AWS region set.');
   else
-    return input({
+    return await input({
       message:
         'Please enter AWS region or use the AWS_REGION environment variable.',
     });