+++ typescript/infra/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hyperlane-xyz/infra
 
+## 3.15.0
+
+### Patch Changes
+
+- Updated dependencies [51bfff683]
+  - @hyperlane-xyz/sdk@3.15.0
+  - @hyperlane-xyz/helloworld@3.15.0
+  - @hyperlane-xyz/utils@3.15.0
+
 ## 3.14.0
 
 ### Patch Changes